LPS-161984 Extract MaxLengthProperties from EditObjectField to its own component
