+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectField/EditObjectField.tsx
@@ -21,16 +21,13 @@ import {
 	Input,
 	InputLocalized,
 	SidePanelForm,
-	Toggle,
 	invalidateRequired,
 	openToast,
 	saveAndReload,
 } from '@liferay/object-js-components-web';
 import {sub} from 'frontend-js-web';
-import React, {useEffect, useRef, useState} from 'react';
-import {createTextMaskInputElement} from 'text-mask-core';
+import React, {useEffect, useState} from 'react';
 
-import {createAutoCorrectedNumberPipe} from '../../utils/createAutoCorrectedNumberPipe';
 import {
 	normalizeFieldSettings,
 	updateFieldSettings,
@@ -44,6 +41,7 @@ import ObjectFieldFormBase, {
 	ObjectFieldErrors,
 	useObjectFieldForm,
 } from '../ObjectFieldFormBase';
+import {MaxLengthProperties} from './MaxLengthProperties';
 import {SearchableContainer} from './SearchableContainer';
 
 import './EditObjectField.scss';
@@ -648,99 +646,6 @@ export default function EditObjectField({
 	);
 }
 
-function MaxLengthProperties({
-	disabled,
-	errors,
-	objectField,
-	objectFieldSettings,
-	onSettingsChange,
-	setValues,
-}: IMaxLengthPropertiesProps) {
-	const [defaultMaxLength, defaultMaxLengthText] =
-		objectField.businessType === 'Text' ? [280, '280'] : [65000, '65,000'];
-
-	const settings = normalizeFieldSettings(objectFieldSettings);
-
-	const inputRef = useRef(null);
-	const maskRef = useRef();
-
-	useEffect(() => {
-		if (settings.showCounter) {
-			maskRef.current = createTextMaskInputElement({
-				guide: false,
-				inputElement: inputRef.current,
-				keepCharPositions: true,
-				mask:
-					objectField.businessType === 'Text'
-						? [/\d/, /\d/, /\d/]
-						: [/\d/, /\d/, /\d/, /\d/, /\d/],
-				pipe: createAutoCorrectedNumberPipe(defaultMaxLength, 1),
-				showMask: true,
-			});
-		}
-	}, [defaultMaxLength, objectField.businessType, settings.showCounter]);
-
-	return (
-		<>
-			<ClayForm.Group>
-				<Toggle
-					disabled={disabled}
-					label={Liferay.Language.get('limit-characters')}
-					name="showCounter"
-					onToggle={(value) => {
-						const updatedSettings: ObjectFieldSetting[] = [
-							{name: 'showCounter', value},
-						];
-
-						if (value) {
-							updatedSettings.push({
-								name: 'maxLength',
-								value: defaultMaxLength,
-							});
-						}
-
-						setValues({objectFieldSettings: updatedSettings});
-					}}
-					toggled={!!settings.showCounter}
-					tooltip={Liferay.Language.get(
-						'when-enabled-a-character-counter-will-be-shown-to-the-user'
-					)}
-				/>
-			</ClayForm.Group>
-			<ClayForm.Group>
-				{settings.showCounter && (
-					<Input
-						disabled={disabled}
-						error={errors.maxLength}
-						feedbackMessage={sub(
-							Liferay.Language.get(
-								'set-the-maximum-number-of-characters-accepted-this-value-cant-be-less-than-x-or-greater-than-x'
-							),
-							'1',
-							defaultMaxLengthText
-						)}
-						label={Liferay.Language.get(
-							'maximum-number-of-characters'
-						)}
-						onChange={({target: {value}}) =>
-							onSettingsChange({
-								name: 'maxLength',
-								value: value && Number(value),
-							})
-						}
-						onInput={({target: {value}}: any) =>
-							(maskRef.current as any).update(value)
-						}
-						ref={inputRef}
-						required
-						value={`${settings.maxLength}`}
-					/>
-				)}
-			</ClayForm.Group>
-		</>
-	);
-}
-
 function AttachmentProperties({
 	errors,
 	objectFieldSettings,
@@ -828,15 +733,6 @@ interface IAttachmentPropertiesProps {
 	errors: ObjectFieldErrors;
 	objectFieldSettings: ObjectFieldSetting[];
 	onSettingsChange: (setting: ObjectFieldSetting) => void;
-}
-
-interface IMaxLengthPropertiesProps {
-	disabled?: boolean;
-	errors: ObjectFieldErrors;
-	objectField: Partial<ObjectField>;
-	objectFieldSettings: ObjectFieldSetting[];
-	onSettingsChange: (setting: ObjectFieldSetting) => void;
-	setValues: (values: Partial<ObjectField>) => void;
 }
 
 interface IProps {
